+++ src/Helper/FormHelperChoiceRadio.php
@@ -33,10 +33,10 @@ class FormHelperChoiceRadio extends AbstractFormHelperChoice
             $htmlElement->setChecked(true);
         }
 
-        foreach ($this->getOptions()->get() as $key => $value) {
+        foreach ($this->getOptions()->get() as $attributeName => $attributeValue) {
             // prevent, that already set attributes are overwritten by options
-            if (!$htmlElement->hasAttribute($key)) {
-                $htmlElement->setAttribute($key, $value);
+            if (!$htmlElement->hasAttribute($attributeName) || $this->isAttributeInjectable($attributeName)) {
+                $htmlElement->setAttribute($attributeName, $attributeValue);
             }
         }