+++ src/Helper/FormHelperChoiceSelect.php
@@ -50,10 +50,10 @@ class FormHelperChoiceSelect extends AbstractFormHelperChoice
             $htmlElement->addChildElement($htmlChildElement);
         }
 
-        foreach ($this->getOptions()->get() as $key => $value) {
+        foreach ($this->getOptions()->get() as $attributeName => $attributeValue) {
             // prevent, that already set attributes are overwritten by options
-            if (!$htmlElement->hasAttribute($key)) {
-                $htmlElement->setAttribute($key, $value);
+            if (!$htmlElement->hasAttribute($attributeName) || $this->isAttributeInjectable($attributeName)) {
+                $htmlElement->setAttribute($attributeName, $attributeValue);
             }
         }