+++ src/Helper/FormHelperFieldHidden.php
@@ -28,10 +28,10 @@ class FormHelperFieldHidden extends AbstractFormHelperField
     {
         $htmlElement = new HtmlInputHidden($this->getProperty()->getFormName(), $this->getProperty()->getFormValue());
 
-        foreach ($this->getOptions()->get() as $key => $value) {
+        foreach ($this->getOptions()->get() as $attributeName => $attributeValue) {
             // prevent, that already set attributes are overwritten by options
-            if (!$htmlElement->hasAttribute($key)) {
-                $htmlElement->setAttribute($key, $value);
+            if (!$htmlElement->hasAttribute($attributeName) || $this->isAttributeInjectable($attributeName)) {
+                $htmlElement->setAttribute($attributeName, $attributeValue);
             }
         }