+++ tests/Helper/FormHelperChoiceRadioTest.php
@@ -9,40 +9,91 @@
  */
 namespace Naucon\Form\Tests;
 
+use Naucon\Form\FormCollectionInterface;
+use Naucon\Form\FormInterface;
 use Naucon\Form\Helper\FormHelperChoiceRadio;
+use Naucon\Form\Mapper\EntityContainerInterface;
 use Naucon\Form\Mapper\Property;
 use Naucon\Form\Tests\Entities\User;
+use Naucon\Utility\Map;
 
+/**
+ * Class FormHelperChoiceRadioTest
+ *
+ * @package Naucon\Form\Tests
+ */
 class FormHelperChoiceRadioTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var FormInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $form;
+
+    /**
+     * @var FormCollectionInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $formCollection;
+
+    /**
+     * @var EntityContainerInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $entityContainer;
+
+    /**
+     * @var Map|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $entityContainerMap;
+
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->form = $this->getMockBuilder(FormInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->formCollection = $this->getMockBuilder(FormCollectionInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->entityContainer = $this->getMockBuilder(EntityContainerInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->entityContainerMap = $this->getMockBuilder(Map::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+    }
+
+
     public function testInit()
     {
         $userEntity = new User();
         $userEntity->setComment('foo');
 
-        $entityContainerMap = $this->getMock('Naucon\Utility\Map');
-        $entityContainerMap->expects($this->any())
+        $this->entityContainerMap->expects($this->any())
             ->method('count')
             ->will($this->returnValue(1));
 
-        $form = $this->getMock('Naucon\Form\FormInterface');
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getName')
             ->will($this->returnValue('testform'));
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getEntityContainerMap')
-            ->will($this->returnValue($entityContainerMap));
+            ->will($this->returnValue($this->entityContainerMap));
 
-        $entityContainer = $this->getMock('Naucon\Form\Mapper\EntityContainerInterface');
-        $entityContainer->expects($this->any())
+
+        $this->entityContainer->expects($this->any())
             ->method('getForm')
-            ->will($this->returnValue($form));
+            ->will($this->returnValue($this->form));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getEntity')
             ->will($this->returnValue($userEntity));
 
-        $propertyObject = new Property($entityContainer, 'comment');
+        $propertyObject = new Property($this->entityContainer, 'comment');
 
         $formHelper = new FormHelperChoiceRadio();
         $formHelper->setProperty($propertyObject);
@@ -52,38 +103,79 @@ class FormHelperChoiceRadioTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('<input type="radio" name="testform[comment]" value="foo" id="testform_comment" checked="checked" />' . PHP_EOL, $formHelper->render());
     }
 
+    public function testRenderWithWhitelistedAttributes()
+    {
+        $userEntity = new User();
+        $userEntity->setComment('foo');
+
+        $this->entityContainerMap->expects($this->any())
+            ->method('count')
+            ->will($this->returnValue(1));
+
+
+        $this->form->expects($this->any())
+            ->method('getName')
+            ->will($this->returnValue('testform'));
+
+        $this->form->expects($this->any())
+            ->method('getEntityContainerMap')
+            ->will($this->returnValue($this->entityContainerMap));
+
+
+        $this->entityContainer->expects($this->any())
+            ->method('getForm')
+            ->will($this->returnValue($this->form));
+
+        $this->entityContainer->expects($this->any())
+            ->method('getEntity')
+            ->will($this->returnValue($userEntity));
+
+        $propertyObject = new Property($this->entityContainer, 'comment');
+
+        $formHelper = new FormHelperChoiceRadio();
+        $formHelper->setProperty($propertyObject);
+        $formHelper->setChoice('foo');
+        $formHelper->setOptions([
+            'id' => 'foo2',
+            'class' => 'form-control'
+        ]);
+        $formHelper->render();
+
+        $this->assertEquals('<input type="radio" name="testform[comment]" value="foo" id="foo2" checked="checked" class="form-control" />' . PHP_EOL, $formHelper->render());
+    }
+
     public function testHelperWithOneFormCollection()
     {
         $userEntity = new User();
         $userEntity->setComment('foo');
 
-        $entityContainerMap = $this->getMock('Naucon\Utility\Map');
-        $entityContainerMap->expects($this->any())
+        $this->entityContainerMap->expects($this->any())
             ->method('count')
             ->will($this->returnValue(1));
 
-        $form = $this->getMock('Naucon\Form\FormCollectionInterface');
-        $form->expects($this->any())
+
+        $this->formCollection->expects($this->any())
             ->method('getName')
             ->will($this->returnValue('testform'));
-        $form->expects($this->any())
+
+        $this->formCollection->expects($this->any())
             ->method('getEntityContainerMap')
-            ->will($this->returnValue($entityContainerMap));
+            ->will($this->returnValue($this->entityContainerMap));
 
-        $entityContainer = $this->getMock('Naucon\Form\Mapper\EntityContainerInterface');
-        $entityContainer->expects($this->any())
+
+        $this->entityContainer->expects($this->any())
             ->method('getForm')
-            ->will($this->returnValue($form));
+            ->will($this->returnValue($this->formCollection));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getName')
             ->will($this->returnValue(0));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getEntity')
             ->will($this->returnValue($userEntity));
 
-        $propertyObject = new Property($entityContainer, 'comment');
+        $propertyObject = new Property($this->entityContainer, 'comment');
 
         $formHelper = new FormHelperChoiceRadio();
         $formHelper->setProperty($propertyObject);