+++ tests/Helper/FormHelperFieldTextTest.php
@@ -9,40 +9,90 @@
  */
 namespace Naucon\Form\Tests;
 
+use Naucon\Form\FormCollectionInterface;
+use Naucon\Form\FormInterface;
 use Naucon\Form\Helper\FormHelperFieldText;
+use Naucon\Form\Mapper\EntityContainerInterface;
 use Naucon\Form\Mapper\Property;
 use Naucon\Form\Tests\Entities\User;
+use Naucon\Utility\Map;
 
+/**
+ * Class FormHelperFieldTextTest
+ *
+ * @package Naucon\Form\Tests
+ */
 class FormHelperFieldTextTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var FormInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $form;
+
+    /**
+     * @var FormCollectionInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $formCollection;
+
+    /**
+     * @var EntityContainerInterface|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $entityContainer;
+
+    /**
+     * @var Map|\PHPUnit_Framework_MockObject_MockObject
+     */
+    protected $entityContainerMap;
+
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->form = $this->getMockBuilder(FormInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->formCollection = $this->getMockBuilder(FormCollectionInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->entityContainer = $this->getMockBuilder(EntityContainerInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $this->entityContainerMap = $this->getMockBuilder(Map::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+    }
+
     public function testInit()
     {
         $userEntity = new User();
         $userEntity->setComment('foo');
 
-        $entityContainerMap = $this->getMock('Naucon\Utility\Map');
-        $entityContainerMap->expects($this->any())
+        $this->entityContainerMap->expects($this->any())
             ->method('count')
             ->will($this->returnValue(1));
 
-        $form = $this->getMock('Naucon\Form\FormInterface');
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getName')
             ->will($this->returnValue('testform'));
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getEntityContainerMap')
-            ->will($this->returnValue($entityContainerMap));
+            ->will($this->returnValue($this->entityContainerMap));
 
-        $entityContainer = $this->getMock('Naucon\Form\Mapper\EntityContainerInterface');
-        $entityContainer->expects($this->any())
+
+        $this->entityContainer->expects($this->any())
             ->method('getForm')
-            ->will($this->returnValue($form));
+            ->will($this->returnValue($this->form));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getEntity')
             ->will($this->returnValue($userEntity));
 
-        $propertyObject = new Property($entityContainer, 'comment');
+        $propertyObject = new Property($this->entityContainer, 'comment');
 
         $formHelper = new FormHelperFieldText();
         $formHelper->setProperty($propertyObject);
@@ -56,33 +106,33 @@ class FormHelperFieldTextTest extends \PHPUnit_Framework_TestCase
         $userEntity = new User();
         $userEntity->setComment('foo');
 
-        $entityContainerMap = $this->getMock('Naucon\Utility\Map');
-        $entityContainerMap->expects($this->any())
+        $this->entityContainerMap->expects($this->any())
             ->method('count')
             ->will($this->returnValue(1));
 
-        $form = $this->getMock('Naucon\Form\FormCollectionInterface');
-        $form->expects($this->any())
+
+        $this->formCollection->expects($this->any())
             ->method('getName')
             ->will($this->returnValue('testform'));
-        $form->expects($this->any())
+
+        $this->formCollection->expects($this->any())
             ->method('getEntityContainerMap')
-            ->will($this->returnValue($entityContainerMap));
+            ->will($this->returnValue($this->entityContainerMap));
 
-        $entityContainer = $this->getMock('Naucon\Form\Mapper\EntityContainerInterface');
-        $entityContainer->expects($this->any())
+
+        $this->entityContainer->expects($this->any())
             ->method('getForm')
-            ->will($this->returnValue($form));
+            ->will($this->returnValue($this->formCollection));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getName')
             ->will($this->returnValue(0));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getEntity')
             ->will($this->returnValue($userEntity));
 
-        $propertyObject = new Property($entityContainer, 'comment');
+        $propertyObject = new Property($this->entityContainer, 'comment');
 
         $formHelper = new FormHelperFieldText();
         $formHelper->setProperty($propertyObject);
@@ -96,34 +146,75 @@ class FormHelperFieldTextTest extends \PHPUnit_Framework_TestCase
         $userEntity = new User();
         $userEntity->setComment('">&Römer');
 
-        $entityContainerMap = $this->getMock('Naucon\Utility\Map');
-        $entityContainerMap->expects($this->any())
+        $this->entityContainerMap->expects($this->any())
             ->method('count')
             ->will($this->returnValue(1));
 
-        $form = $this->getMock('Naucon\Form\FormInterface');
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getName')
             ->will($this->returnValue('testform'));
-        $form->expects($this->any())
+
+        $this->form->expects($this->any())
             ->method('getEntityContainerMap')
-            ->will($this->returnValue($entityContainerMap));
+            ->will($this->returnValue($this->entityContainerMap));
 
-        $entityContainer = $this->getMock('Naucon\Form\Mapper\EntityContainerInterface');
-        $entityContainer->expects($this->any())
+
+        $this->entityContainer->expects($this->any())
             ->method('getForm')
-            ->will($this->returnValue($form));
+            ->will($this->returnValue($this->form));
 
-        $entityContainer->expects($this->any())
+        $this->entityContainer->expects($this->any())
             ->method('getEntity')
             ->will($this->returnValue($userEntity));
 
-        $propertyObject = new Property($entityContainer, 'comment');
+        $propertyObject = new Property($this->entityContainer, 'comment');
 
         $formHelper = new FormHelperFieldText();
         $formHelper->setProperty($propertyObject);
         $formHelper->render();
 
         $this->assertEquals('<input type="text" name="testform[comment]" value="&#34;>&#38;Römer" id="testform_comment" />' . PHP_EOL, $formHelper->render());
+    }
+
+
+    public function testRenderWithWhitelistedAttributes()
+    {
+        $userEntity = new User();
+        $userEntity->setComment('foo');
+
+        $this->entityContainerMap->expects($this->any())
+            ->method('count')
+            ->will($this->returnValue(1));
+
+
+        $this->form->expects($this->any())
+            ->method('getName')
+            ->will($this->returnValue('testform'));
+
+        $this->form->expects($this->any())
+            ->method('getEntityContainerMap')
+            ->will($this->returnValue($this->entityContainerMap));
+
+
+        $this->entityContainer->expects($this->any())
+            ->method('getForm')
+            ->will($this->returnValue($this->form));
+
+        $this->entityContainer->expects($this->any())
+            ->method('getEntity')
+            ->will($this->returnValue($userEntity));
+
+        $propertyObject = new Property($this->entityContainer, 'comment');
+
+        $formHelper = new FormHelperFieldText();
+        $formHelper->setProperty($propertyObject);
+        $formHelper->setOptions([
+            'id' => 'foo2',
+            'class' => 'form-control'
+        ]);
+        $formHelper->render();
+
+        $this->assertEquals('<input type="text" name="testform[comment]" value="foo" id="foo2" class="form-control" />' . PHP_EOL, $formHelper->render());
     }
 }